Added Gemini first pass of application with following prompt
You're an expert in visualising audio, typescript and threejs.

I want to make an application that shows what notes are being played. So the application needs to take in input from the microphone and put it through a fourier transform to extract the amplitude of the different frequencies. I would then like to display these frequency amplitudes on a 2D plot that moves from right to left, i.e. it always shows the current frequency response on the right side and has the past frequency amplitudes scrolling off to the left. The higher the amplitude of the frequency, the brighter the canvas should be. Please provide a simple colour scale for now going from black to white.

Underneath the canvas showing the frequency amplitude response I would like to show the dominant keys being played such as C3, A5# (sharp), G4f (flat), etc. These should show as a column of text under the canvas as it canvas scrolls from right to left. For now just limit it to the top 5 frequencies above some reasonable amplitude threshold. Also there should be a minimum time needed to be counted as a key, perhaps 1/20th of a second? So 50 milliseconds? And if there are lots of different keys it's ok for now if these elements overlap. Later we can put an on hover to have their Z-index set to bring them to the front, or have the ability to expand the time scale / zoom in on the time line to make it stretch out so that the different notes can be seen easily. Thank you.