Some fixes
Used the prompt:

Can you edit this please to:
1. Allow for the maximum frequency shown to be set as a constant (and to reduce it to half of what it currently is)
2. Replace the notes-container canvas with a list of html elements.  Also if there is a note that is present from one moment to the next then it is not necessary to show it again.  For example instead of showing:
```
| C4 | G5 | C4 |
|    | F5 |    |
|    | E5 |    |
|    | D5 |    |
|    | C4 |    |
```
Then only this should be shown.  Here you can see that C4 is still in the top 5 list of loudest notes so keep showing it, don't displace it by the G5 which is actually currently louder than it.
```
| C4 |    |    |
|    | G5 |    |
|    | F5 |    |
|    | E5 |    |
|    | D5 |    |
```

And if there are more than the NOTES_TO_SHOW notes detected, and then less than that number then even then a current note (C4 in this example) should not be redisplayed.  In this example G5, F5, E5, D5, and B5 are all louder than C4 but only last for one moment, where as C4 sustains longer than them.

```
| C4 | G5 |    |
|    | F5 |    |
|    | E5 |    |
|    | D5 |    |
|    | B5 |    |
```