Refactor and Attempt at recording and replaying essence of the sound
First prompt to Gemini 3.0 Pro (Preview) broke the application:

I would like you to add another file with some code which can be used to
instantiate a recorder. This should then be used to record up to NOTES_TO_RECORD
for up to RECORD_DURATION_MS.
At the end of RECORD_DURATION_MS there should be an option to replace the
recorded notes. I would like you to add a third file which will contain the
logic to take the recording from the second file and play it back as simple
tones of the correct frequencies and durations.

Second prompt to Gemini 3.0 Pro (Preview) (including spelling mistake):

Your last changes to demoSim were broken and unacceptable.  Please refactor
this file into several smaller files so that you can better reason and work
on this please.